+++ src/pages/cart/cart.js
@@ -1,52 +1,10 @@
-/* ---------- toggle ---------- */
-const cartListButton = document.querySelectorAll('.cart__item > button');
+import '@/styles/style.scss';
 
-cartListButton.forEach((button) => {
-  button.addEventListener('click', function () {
-    const selectedList = this.nextElementSibling;
-    const arrowButton = this.querySelector('.icon--arrow--bottom');
-    const selectedItem = selectedList.querySelectorAll('.selected__item');
+import '@/layout/header/header.js';
+import '@/layout/footer/footer.js';
 
-    if (selectedList.children.length !== 0) {
-      selectedList.classList.toggle('is--open');
-      arrowButton.classList.toggle('is--open');
-    } else {
-      selectedList.classList.remove('is--open');
-      arrowButton.classList.remove('is--open');
-    }
-
-    selectedItem.forEach((liELement) => {
-      const deleteButton = liELement.querySelector('.delete');
-
-      deleteButton.addEventListener('click', function () {
-        this.parentElement.remove();
-
-        if (selectedList.children.length !== 0) {
-          selectedList.classList.add('is--open');
-          arrowButton.classList.add('is--open');
-          console.log(selectedList.children.length);
-        } else {
-          console.log(selectedList.children.length);
-          selectedList.classList.remove('is--open');
-          arrowButton.classList.remove('is--open');
-        }
-      });
-    });
-  });
-});
-
-/* ---------- checkbox ---------- */
-const checkboxAll = document.querySelector('#checkbox02');
-const checkboxSubAll = document.querySelectorAll('.selected__item .checkbox');
-
-checkboxAll.addEventListener('click', function () {
-  const isChecked = checkboxAll.checked;
-
-  checkboxSubAll.forEach((checkbox) => {
-    if (isChecked) {
-      checkbox.checked = true;
-    } else {
-      checkbox.checked = false;
-    }
-  });
-});
+import '@/pages/cart/address.js';
+import '@/pages/cart/delete.js';
+import '@/pages/cart/order.js';
+import '@/pages/cart/toggle.js';
+import '@/pages/cart/checkbox.js';
