+++ src/pages/cart/order.js
@@ -1,21 +1,21 @@
-import modalHandle from '@/lib/modal';
+import modalHandle from '@/lib/modal.js';
+import timerFn from '@/lib/timer.js';
 
 const orderButton = document.querySelector('.button--order');
 const orderConfirm = document.querySelector('.button--confirm');
 
 /* ---------- 주문하기 버튼 클릭 시 ---------- */
 orderButton.addEventListener('click', function () {
-  // timer();
-
-  modalHandle('.modal__order', 'is--open', open);
+  modalHandle('.modal', 'is--open', open);
+  timerFn();
 
   setTimeout(() => {
-    modalHandle('.modal__order', 'is--open', close);
+    modalHandle('.modal', 'is--open', close);
     location.href = '/';
   }, 5000);
 });
 
 orderConfirm.addEventListener('click', function () {
-  modalHandle('.modal__order', 'is--open', close);
+  modalHandle('.modal', 'is--open', close);
   location.href = '/';
 });
