+++ src/styles/components/_address.scss
@@ -59,5 +59,10 @@
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
+    display: none;
+
+    &.is--open {
+      display: block;
+    }
   }
 }
